requirements.txt
streamlit==1.30.0
numpy==1.26.4
pandas==2.1.4
opencv-python-headless==4.8.1.78
pillow==9.5.0
streamlit-option-menu==0.3.6
easyocr==1.7.0

